+++ js/photo-sources.js
@@ -1,0 +1,20 @@
+const PHOTO_COUNTER = 25;
+const MESSAGES = ['Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'В целом всё неплохо. Но не всё.'];
+const NAMES = [
+  'Иван',
+  'Cтепан',
+  'Василий',
+  'Йосиф',
+  'Владимир',
+  'Джордж'
+];
+const DESCRIPTIONS = [
+  'Хорошее фото',
+  'Идеально фото'
+];
+
+export {PHOTO_COUNTER,
+  NAMES,
+  DESCRIPTIONS,
+  MESSAGES
+};
